[clients_gui] fix signed/unsigned warning in LocationEditorToolbar

+++ src/libraries/clients_gui/LocationEditorToolbar.h
@@ -54,7 +54,7 @@ private slots:
     void Goto();
     void AddParamPoint();
     void CreateBookmark();
-    void GotoBookmark( int index );
+    void GotoBookmark( size_t index );
     void ClearBookmarks();
     //@}
